fix: move PingExample to its own class
